move trap.rs into executor submodule
