+++ crates/wasmi/src/engine/regmach/executor/trap.rs
